requirements.txt
fastapi==0.120.3
ipython==9.5.0
pandas==2.3.3
pubchempy==1.0.5
pwdlib==0.3.0
pydantic==2.12.3
pydantic_settings==2.11.0
pytest==8.4.2
python_jose==3.5.0
rdkit==2025.3.6
Requests==2.32.5
SQLAlchemy==2.0.44
uvicorn==0.38.0
pydantic-settings
